requirements.txt
# ====================================
# ATS Resume Analyzer v3.0
# Core Dependencies with Pinned Versions
# ====================================

# --- Web Framework & UI ---
streamlit==1.28.1              # Web framework for interactive apps
streamlit-option-menu==0.3.6   # Custom menu component
altair==5.1.2                  # Advanced visualization grammar

# --- Data Processing & Analysis ---
pandas==2.1.3                  # Data manipulation and analysis
numpy==1.26.2                  # Numerical computing
scikit-learn==1.3.2            # Machine learning algorithms
scipy==1.11.4                  # Scientific computing library

# --- PDF & Document Processing ---
PyPDF2==4.0.1                  # PDF extraction and manipulation
python-docx==0.8.11            # Read/write Word documents

# --- Visualization & Charts ---
plotly==5.18.0                 # Interactive visualizations
matplotlib==3.8.2              # Static plotting library
seaborn==0.13.0                # Statistical data visualization

# --- Authentication & Security ---
bcrypt==4.1.1                  # Password hashing (future enhancement)
cryptography==41.0.7           # Cryptographic recipes

# --- Text Processing & NLP ---
nltk==3.8.1                    # Natural language processing toolkit
textblob==0.17.1               # Simplified text NLP

# --- Utilities ---
python-dotenv==1.0.0           # Environment variable management
requests==2.31.0               # HTTP library for API calls
Pillow==10.1.0                 # Image processing library

# --- Development & Testing (optional) ---
# pytest==7.4.3                # Unit testing framework
# black==23.12.1               # Code formatter
# flake8==6.1.0                # Linter
# pylint==3.0.3                # Code analysis

# ====================================
# Installation Instructions:
# ====================================
# pip install -r requirements.txt
#
# For development with optional tools:
# pip install -r requirements.txt
# pip install pytest black flake8 pylint
#
# ====================================
# Version Notes:
# ====================================
# Python: 3.8+ (tested with 3.13.8)
# All packages pinned to stable releases
# as of February 2025
# ====================================